Move TransportLayer and NodeDiscovery implementations moved to comm
