+++ comm/src/main/scala/coop/rchain/comm/transport/SslSessionClientInterceptor.scala
@@ -1,8 +1,7 @@
-package coop.rchain.node.effects
+package coop.rchain.comm.transport
 
 import coop.rchain.comm.protocol.routing._
 import coop.rchain.comm.protocol.routing.TLResponse.Payload
-import coop.rchain.node.CertificateHelper
 
 import io.grpc._
 import javax.net.ssl.SSLSession
